+++ heylogs-api/src/main/java/nbbrd/heylogs/spi/CompareLink.java
@@ -1,0 +1,18 @@
+package nbbrd.heylogs.spi;
+
+import lombok.NonNull;
+
+import java.net.URL;
+
+public interface CompareLink extends ForgeLink {
+
+    @NonNull
+    URL toURL();
+
+    @NonNull
+    CompareLink derive(@NonNull String tag);
+
+    @NonNull
+    URL getProjectURL();
+
+}