+++ heylogs-ext-forgejo/src/test/java/nbbrd/heylogs/ext/forgejo/ForgejoTest.java
@@ -1,4 +1,4 @@
-package nbbrd.heylogs.ext.github;
+package nbbrd.heylogs.ext.forgejo;
 
 import nbbrd.heylogs.spi.Forge;
 import org.junit.jupiter.api.Test;
@@ -8,29 +8,29 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 import static tests.heylogs.spi.ForgeAssert.assertForgeCompliance;
 
-class GitHubTest {
+class ForgejoTest {
 
     @Test
     void testCompliance() {
-        assertForgeCompliance(new GitHub());
+        assertForgeCompliance(new Forgejo());
     }
 
     @Test
     void testIsCompareLink() {
-        Forge x = new GitHub();
+        Forge x = new Forgejo();
         assertThat(x.isCompareLink(urlOf("https://nbb.be"))).isFalse();
-        assertThat(x.isCompareLink(urlOf("https://github.com/nbbrd/heylogs/compare/v0.7.2...HEAD"))).isTrue();
-        assertThat(x.isCompareLink(urlOf("https://localhost:8080/nbbrd/heylogs/compare/v0.7.2...HEAD"))).isFalse();
+        assertThat(x.isCompareLink(urlOf("https://codeberg.org/Freeyourgadget/Gadgetbridge/compare/0.86.0...0.86.1"))).isTrue();
+        assertThat(x.isCompareLink(urlOf("https://localhost:8080/Freeyourgadget/Gadgetbridge/compare/0.86.0...0.86.1"))).isFalse();
     }
 
     @Test
     void testGetProjectURL() {
-        Forge x = new GitHub();
+        Forge x = new Forgejo();
 
         assertThatIllegalArgumentException()
                 .isThrownBy(() -> x.getProjectURL(urlOf("https://nbb.be")));
 
-        assertThat(x.getProjectURL(urlOf("https://github.com/nbbrd/heylogs/compare/v0.7.2...HEAD")))
-                .isEqualTo(urlOf("https://github.com/nbbrd/heylogs"));
+        assertThat(x.getProjectURL(urlOf("https://codeberg.org/Freeyourgadget/Gadgetbridge/compare/0.86.0...0.86.1")))
+                .isEqualTo(urlOf("https://codeberg.org/Freeyourgadget/Gadgetbridge"));
     }
 }